IntakeSubsystem for Coral
Once the command is called, it the motors will continuously spin to redirect the coral into the slot until the sensor receives a signal to stop.
It also works in reverse, if it the button was pressed, and it received coral, the next time the command is called the motors spin in reverse until the coral is removed.